+++ Modules/Countries/Http/Controllers/CountryController.php
@@ -31,22 +31,22 @@ class CountryController extends Controller
     {
         try {
             if ($request->has('onlyTrashed') && $request->onlyTrashed) {
-                $records = Country::onlyTrashed()->latest()->paginate(9);
+                $records = Country::onlyTrashed()->latest()->paginate(config('setting.paginate'));
             } else {
-                $records = Country::latest()->paginate(9);
+                $records = Country::latest()->paginate(config('setting.paginate'));
             }
             if ($records->count() > 0) {
                 return response()->json(
                     [
                         'message' => __('Returned Successfully'),
                         'data' => CountryResource::collection($records)->response()->getData(true)
-                    ],200);
+                    ], 200);
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => $request],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => $request], 400);
         }
     }
 
@@ -66,13 +66,13 @@ class CountryController extends Controller
                 $store->save();
             }
             if ($store) {
-                return response()->json(['message' => __('Updated Successfully'), 'data' => ''],200);
+                return response()->json(['message' => __('Inserted Successfully'), 'data' => ''], 200);
             } else {
-                return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+                return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => $request],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => $request], 400);
         }
     }
 
@@ -86,13 +86,13 @@ class CountryController extends Controller
         try {
             $record = Country::find($id);
             if ($record) {
-                return response()->json(['message' => __('Changed Successfully'), 'data' => CountryResource::make($record)],200);
+                return response()->json(['message' => __('Changed Successfully'), 'data' => CountryResource::make($record)], 200);
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
         }
     }
 
@@ -115,16 +115,16 @@ class CountryController extends Controller
                     $record->save();
                 }
                 if ($update) {
-                    return response()->json(['message' => __('Updated Successfully'), 'data' => ''],200);
+                    return response()->json(['message' => __('Updated Successfully'), 'data' => ''], 200);
                 } else {
-                    return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+                    return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
                 }
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => $request],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => $request], 400);
         }
     }
 
@@ -138,18 +138,24 @@ class CountryController extends Controller
         try {
             $record = Country::find($id);
             if ($record) {
+//                $result = checkRelation($record, ['clients', 'sellers']);
+                $result = 0;
+                if ($result) {
+                    return response()->json(['message' => __('You can not delete this record'), 'data' => ''], 400);
+                }
+
                 $del = $record->delete();
                 if ($del) {
-                    return response()->json(['message' => __('Changed Successfully'), 'data' => ''],200);
+                    return response()->json(['message' => __('Deleted Successfully'), 'data' => ''], 200);
                 } else {
-                    return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+                    return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
                 }
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
         }
     }
 
@@ -165,16 +171,16 @@ class CountryController extends Controller
             if ($record) {
                 $restore = $record->restore();
                 if ($restore) {
-                    return response()->json(['message' => __('Restored Successfully'), 'data' => ''],200);
+                    return response()->json(['message' => __('Restored Successfully'), 'data' => ''], 200);
                 } else {
-                    return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+                    return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
                 }
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
         }
     }
 
@@ -188,18 +194,24 @@ class CountryController extends Controller
         try {
             $record = Country::find($id);
             if ($record) {
+//                $result = checkRelation($record, ['clients', 'sellers']);
+                $result = 0;
+                if ($result) {
+                    return response()->json(['message' => __('You can not delete this record'), 'data' => ''], 400);
+                }
+
                 $del = $record->forceDelete();
                 if ($del) {
-                    return response()->json(['message' => __('Force Deleted Successfully'), 'data' => ''],200);
+                    return response()->json(['message' => __('Force Deleted Successfully'), 'data' => ''], 200);
                 } else {
-                    return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+                    return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
                 }
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => ''],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => ''], 400);
         }
     }
 
@@ -216,13 +228,13 @@ class CountryController extends Controller
                 $record->update([
                     'active' => $request->status
                 ]);
-                return response()->json(['message' => __('Changed Successfully'), 'data' => ''],200);
+                return response()->json(['message' => __('Changed Successfully'), 'data' => ''], 200);
             } else {
-                return response()->json(['message' => __('Model not found'), 'data' => ''],400);
+                return response()->json(['message' => __('Model not found'), 'data' => ''], 400);
             }
         } catch (Throwable $e) {
             Log::error($e);
-            return response()->json(['message' => __('Something went wrong'), 'data' => $request],400);
+            return response()->json(['message' => __('Something went wrong'), 'data' => $request], 400);
         }
     }
 }
